+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3 (unreleased)
+
+- Added `maxDepth` option
+- Fixed error with nested context
+
 ## 0.1.2 (2023-02-11)
 
 - Fixed warnings with PHP 8.2